+++ src/main/java/proguard/classfile/VersionConstants.java
@@ -58,6 +58,8 @@ public class VersionConstants
     public static final int CLASS_VERSION_15_MINOR  = 0;
     public static final int CLASS_VERSION_16_MAJOR  = 60;
     public static final int CLASS_VERSION_16_MINOR  = 65535; // Includes preview versions.
+    public static final int CLASS_VERSION_17_MAJOR  = 61;
+    public static final int CLASS_VERSION_17_MINOR  = 0;
 
     public static final int CLASS_VERSION_1_0 = (CLASS_VERSION_1_0_MAJOR << 16) | CLASS_VERSION_1_0_MINOR;
     public static final int CLASS_VERSION_1_2 = (CLASS_VERSION_1_2_MAJOR << 16) | CLASS_VERSION_1_2_MINOR;
@@ -75,4 +77,7 @@ public class VersionConstants
     public static final int CLASS_VERSION_14  = (CLASS_VERSION_14_MAJOR  << 16) | CLASS_VERSION_14_MINOR;
     public static final int CLASS_VERSION_15  = (CLASS_VERSION_15_MAJOR  << 16) | CLASS_VERSION_15_MINOR;
     public static final int CLASS_VERSION_16  = (CLASS_VERSION_16_MAJOR  << 16) | CLASS_VERSION_16_MINOR;
+    public static final int CLASS_VERSION_17  = (CLASS_VERSION_17_MAJOR  << 16) | CLASS_VERSION_17_MINOR;
+
+    public static final int MAX_SUPPORTED_VERSION = CLASS_VERSION_17;
 }